Introduced DiscoveredResource abstraction to make resource discovery plugable.
Extracted common code of the store system resource discovery to make it usable in a standalone way. Customer app now defines a DiscoveredResource that's based on a ServiceInstanceProvider and a traversal. The former by default is a statically defined other system (localhost:8081). The latter is a simple traversal of three rels. 

Also extracted a DiscoveredResourceRefresher that will verify the discovered resource in a configured default time interval. Added Spring Boot auto-configuration to read these timing settings from application properties and default them to a verification every 5 seconds with an initial delay of 10 seconds.

In the "cloud" profile we enable service discovery via Eureka an dynamically lookup a "stores" service.